+++ src/window/window_temple_complex_info.cpp
@@ -17,7 +17,7 @@ void temple_complex_info_window::init(object_info &c) {
 
     auto temple = smart_cast<building_temple_complex>(c.building_get());
 
-    int image_offset = 0;
+    int16_t image_offset = 0;
     switch (temple->type()) {
     case BUILDING_TEMPLE_COMPLEX_OSIRIS: image_offset = 21; break;
     case BUILDING_TEMPLE_COMPLEX_RA: image_offset = 22; break;
